Make k8 deployers use functionalUnitForGoal
