+++ src/software-delivery-machine/blueprint/deploy/k8sDeploy.ts
@@ -25,27 +25,20 @@ import { NoticeK8sTestDeployCompletionOnStatus } from "../../../handlers/events/
 import { NoticeK8sProdDeployCompletionOnStatus } from "../../../handlers/events/delivery/deploy/k8s/NoticeK8sProdDeployCompletion";
 import { requestDeployToK8s } from "../../../handlers/events/delivery/deploy/k8s/RequestK8sDeploys";
 import { ExecuteGoalOnSuccessStatus } from "../../../handlers/events/delivery/ExecuteGoalOnSuccessStatus";
+import { functionalUnitForGoal } from "../../../blueprint/dsl/functionalUnitForGoal";
 
 export const K8sTestingDomain = "testing";
 export const K8sProductionDomain = "production";
 
-export const K8sStagingDeployOnSuccessStatus: FunctionalUnit = {
-    eventHandlers: [
-        () => new ExecuteGoalOnSuccessStatus("K8TestDeploy",
+export const K8sStagingDeployOnSuccessStatus: FunctionalUnit =
+    functionalUnitForGoal("K8TestDeploy",
             StagingDeploymentGoal,
-            requestDeployToK8s(K8sTestingDomain)),
-    ],
-    commandHandlers: [],
-};
+            requestDeployToK8s(K8sTestingDomain));
 
-export const K8sProductionDeployOnSuccessStatus: FunctionalUnit = {
-    eventHandlers: [
-        () => new ExecuteGoalOnSuccessStatus("K8ProductionDeploy",
+export const K8sProductionDeployOnSuccessStatus: FunctionalUnit =
+        functionalUnitForGoal("K8ProductionDeploy",
             ProductionDeploymentGoal,
-            requestDeployToK8s(K8sProductionDomain)),
-    ],
-    commandHandlers: [],
-};
+            requestDeployToK8s(K8sProductionDomain));
 
 export const NoticeK8sTestDeployCompletion = new NoticeK8sTestDeployCompletionOnStatus(
     StagingDeploymentGoal,
